refactor: extract debug routes to routes/debug.js
