+++ backend/server.js
@@ -117,37 +117,7 @@ app.post("/api/fairs/:fairId/refresh-invite-code", verifyFirebaseToken, async (r
 app.use("/api", fairsRouter);
 
 
-app.get("/api/debug/gemini-models", async (req, res) => {
-  try {
-    const key = process.env.GEMINI_API_KEY;
-    const r = await fetch(`https://generativelanguage.googleapis.com/v1beta/models?key=${key}`);
-    const data = await r.json();
-
-    const models = (data.models || [])
-      .filter(m => (m.supportedGenerationMethods || []).includes("generateContent"))
-      .map(m => m.name);
-
-    res.json({ ok: true, models });
-  } catch (e) {
-    res.status(500).json({ ok: false, error: e.message });
-  }
-});
-
-app.get("/api/debug/storage-bucket", async (req, res) => {
-  try {
-    const bucket = admin.storage().bucket(); // ✅ define it here (Option A)
-    const [files] = await bucket.getFiles({ maxResults: 1 });
-
-    return res.json({
-      ok: true,
-      bucket: bucket.name,
-      sampleFile: files?.[0]?.name || null,
-    });
-  } catch (e) {
-    console.error(e);
-    return res.status(500).json({ ok: false, error: e.message });
-  }
-});
+app.use("/api", require("./routes/debug"));
 
 /* ----------------------------------------------------
    STREAM TOKEN ENDPOINT
@@ -4015,20 +3985,6 @@ app.get("/api/booths/:boothId/ratings", verifyFirebaseToken, async (req, res) =>
     return res.status(500).json({ error: "Failed to fetch ratings" });
   }
 });
-
-/* ============================================================
-   PATCH CACHE STATS (DEBUG ENDPOINT)
-============================================================ */
-app.get("/api/debug/patch-cache", verifyFirebaseToken, async (req, res) => {
-  // Only allow admins or in development
-  if (process.env.NODE_ENV !== "development") {
-    return res.status(403).json({ ok: false, error: "Not allowed in production" });
-  }
-
-  const stats = patchCache.getStats();
-  res.json({ ok: true, stats });
-});
-
 
 if (require.main === module) {
   app.listen(PORT, () => {
